# Add guest config and update NodeNumber plugin to use it