--- a/scheduler/plugin/plugin.go
+++ b/scheduler/plugin/plugin.go
@@ -56,7 +56,7 @@ func NewFromConfig(ctx context.Context, config PluginConfig) (framework.Plugin, 
 		return nil, fmt.Errorf("wasm: error reading guest binary at %s: %w", config.GuestPath, err)
 	}
 
-	runtime, guestModule, err := prepareRuntime(ctx, guestBin)
+	runtime, guestModule, err := prepareRuntime(ctx, guestBin, config.GuestConfig)
 	if err != nil {
 		return nil, err
 	}