… examples/nodenumber/main.wasm
… scheduler/test/testdata/error/panic_on_get_config.wasm